+++ src/resources/views/columns/select_multiple.blade.php
@@ -1,7 +1,7 @@
 {{-- relationships with pivot table (n-n) --}}
 <td>
     <?php
-        $results = $entry->{$column['entity']}()->getResults();
+        $results = $entry->{$column['entity']};
 
         if ($results && $results->count()) {
             $results_array = $results->pluck($column['attribute'], 'id');